Make staged debugging scripts for easier demos; also reset template settings in derivation / filter interfaces when the path is changed

+++ docs/exampleDatasets/debugScripts.js
@@ -1,5 +1,14 @@
 /* globals d3, origraph */
-window.prepConnectState = async () => {
+async function checkStage (stage, limit) {
+  if (stage <= limit) {
+    await window.mainView.handleModelChange();
+    return false;
+  } else {
+    return true;
+  }
+}
+
+window.prepMovies = async (stage = Infinity) => {
   const files = [
     'movies/movies.json',
     'movies/credits.json',
@@ -18,9 +27,14 @@ window.prepConnectState = async () => {
     });
     classes[filename] = newClass;
   }
+
+  if (!await checkStage(stage, 0)) { return; }
+
   let movies = classes['movies/movies.json'].interpretAsNodes();
   movies.setClassName('Movies');
   movies.setAnnotation('labelAttr', 'title');
+
+  if (!await checkStage(stage, 1)) { return; }
 
   let cast = classes['movies/credits.json'].unroll('cast')
     .interpretAsEdges();
@@ -32,9 +46,92 @@ window.prepConnectState = async () => {
   crew.setAnnotation('labelAttr', 'job');
   classes['movies/credits.json'].delete();
 
+  if (!await checkStage(stage, 3)) { return; }
+
   let people = classes['movies/people.json'].interpretAsNodes();
   people.setClassName('People');
   people.setAnnotation('labelAttr', 'name');
+
+  if (!await checkStage(stage, 4)) { return; }
+
+  cast.connectToNodeClass({
+    nodeClass: movies,
+    side: 'target',
+    nodeAttribute: 'id',
+    edgeAttribute: 'movie_id'
+  });
+  cast.connectToNodeClass({
+    nodeClass: people,
+    side: 'source',
+    nodeAttribute: 'id',
+    edgeAttribute: 'id'
+  });
+  crew.connectToNodeClass({
+    nodeClass: movies,
+    side: 'target',
+    nodeAttribute: 'id',
+    edgeAttribute: 'movie_id'
+  });
+  crew.connectToNodeClass({
+    nodeClass: people,
+    side: 'source',
+    nodeAttribute: 'id',
+    edgeAttribute: 'id'
+  });
+
+  if (!await checkStage(stage, 5)) { return; }
+
+  let bechdelTests = classes['movies/bechdeltest.json'].interpretAsNodes();
+  bechdelTests.setClassName('Bechdel Tests');
+  bechdelTests.table.deriveAttribute('tt_imdbid', item => 'tt' + item.row.imdbid);
+
+  const hasScore = movies.connectToNodeClass({
+    otherNodeClass: bechdelTests,
+    attribute: 'imdb_id',
+    otherAttribute: 'tt_imdbid'
+  });
+  hasScore.setClassName('Has Score');
+
+  if (!await checkStage(stage, 6)) { return; }
+
+  movies.table.deriveAttribute('Bechdel Score', async (movie) => {
+    for await (const score of movie.edges({ classes: [hasScore] })) {
+      for await (const bechdelTest of score.nodes({ classes: [bechdelTests] })) {
+        return bechdelTest.row['rating'];
+      }
+    }
+    return null;
+  });
+  movies.table.deriveAttribute('Cast Gender Bias', async (movie) => {
+    let nWomen = 0;
+    let nMen = 0;
+    for await (const castEdge of movie.edges({ classes: [cast] })) {
+      for await (const person of castEdge.nodes({ classes: [people] })) {
+        let value = person.row['gender'];
+        if (value === 1) {
+          nWomen++;
+        } else if (value === 2) {
+          nMen++;
+        }
+      }
+    }
+    return nMen / (nWomen + nMen);
+  });
+  movies.table.deriveAttribute('Crew Gender Bias', async (movie) => {
+    let nWomen = 0;
+    let nMen = 0;
+    for await (const crewEdge of movie.edges({ classes: [crew] })) {
+      for await (const person of crewEdge.nodes({ classes: [people] })) {
+        let value = person.row['gender'];
+        if (value === 1) {
+          nWomen++;
+        } else if (value === 2) {
+          nMen++;
+        }
+      }
+    }
+    return nMen / (nWomen + nMen);
+  });
 
   await window.mainView.handleModelChange();
 };
@@ -72,7 +169,7 @@ window.prepParallelEdgesAndNodes = async () => {
   await window.mainView.handleModelChange();
 };
 
-window.halfwayThroughSenate = async () => {
+window.halfwayThroughSenate = async (stage = Infinity) => {
   const files = [
     'twitterPolitics/allContributions.json',
     'twitterPolitics/pressReleases.json',
@@ -95,6 +192,7 @@ window.halfwayThroughSenate = async () => {
   }
 
   // Y1
+  if (!await checkStage(stage, 1)) { return; }
 
   const senators = classes['twitterPolitics/senateMembers.json']
     .interpretAsNodes();
@@ -118,26 +216,30 @@ window.halfwayThroughSenate = async () => {
   votes.setClassName('Votes');
 
   // Y2
+  if (!await checkStage(stage, 2)) { return; }
 
   let twitterAccounts = tweets.expand('user'); // Y3
   const projectedTwitterEdge = Array.from(twitterAccounts.edgeClasses())[0];
 
   // Y4
+  if (!await checkStage(stage, 4)) { return; }
 
   const twitterSenatorLink = twitterAccounts.connectToNodeClass({
     otherNodeClass: senators,
     attribute: 'screen_name',
     otherAttribute: 'twitter_account'
   }); // Y5
+  if (!await checkStage(stage, 5)) { return; }
 
   twitterAccounts = twitterAccounts.interpretAsEdges({ autoconnect: true }); // Y6
+  if (!await checkStage(stage, 6)) { return; }
   projectedTwitterEdge.delete();
   twitterSenatorLink.delete();
-/*
+
   const contribSenatorEdge = contribs.connectToNodeClass({
     otherNodeClass: senators,
-    attribute: 'fec_candidate_id',
-    otherAttribute: 'candidate'
+    attribute: 'candidate',
+    otherAttribute: 'fec_candidate_id'
   });
   senators.connectToNodeClass({
     otherNodeClass: pressReleases,
@@ -146,6 +248,7 @@ window.halfwayThroughSenate = async () => {
   }).setClassName('Released');
 
   // Y7
+  if (!await checkStage(stage, 7)) { return; }
 
   const [ yesses, nos ] = votes
     .closedFacet('vote_position', ['Yes', 'No'])
@@ -165,17 +268,21 @@ window.halfwayThroughSenate = async () => {
   noVotes.setClassName('Voted');
 
   // Y8
+  if (!await checkStage(stage, 8)) { return; }
 
   const donorCommittees = contribs.promote('fec_committee_name'); // Y9
+  if (!await checkStage(stage, 9)) { return; }
   donorCommittees.setClassName('Donor Committees');
 
   const donorContribEdge = Array.from(donorCommittees.edgeClasses())[0];
   contribs = contribs.interpretAsEdges({ autoconnect: true }); // Y10
+  if (!await checkStage(stage, 10)) { return; }
   contribSenatorEdge.delete();
   donorContribEdge.delete();
 
   const [ contribsFor, contribsAgainst ] = contribs
     .closedFacet('support_or_oppose', [ 'S', 'O' ]); // Y11
+  if (!await checkStage(stage, 11)) { return; }
   contribsFor.setClassName('Contributions For');
   contribsAgainst.setClassName('Contributions Against');
   contribs.delete();
@@ -208,8 +315,6 @@ window.halfwayThroughSenate = async () => {
     nos.classId
   ]);
   contribAgainstNo.setClassName('Contributions against No');
-
-  */
 
   // Y12
